Initial implementation of template engine.
This is an initial implementation of an engine for frontend evaluation
of Squarespace templates.

Initially the engine executes instruction trees generated by the Java
compiler. These trees are JSON-encoded S-expressions designed to be
compact, and something it will be trivial for a future JavaScript parser
to build.

The engine also accepts a Visitor which will fire events when certain
instructions are executed. This is intended to allow the existing HTML
annotation feature to be carried over.

Since it can be the provided to the engine at runtime, the annotation
visitor will not be part of this library. This avoids having to weave
the annotation code into the engine statically.

Most of the key instructions, core formatters and predicates, are
implemented with high test coverage.